add and move
add new package for listener and moved it there, added comments for methods
