+++ RaidsPerRegion1.2/bin/.gitignore
@@ -1,0 +1,1 @@
+/me/
